+++ apps/ui-dev/component/feedback/alert/MainView.jsx
@@ -1,0 +1,37 @@
+import EditFilled from "@elf-framework/icon/EditFilled";
+import InfoOutlined from "@elf-framework/icon/InfoOutlined";
+import { Alert, Flex, VBox } from "@elf-framework/ui";
+
+export function MainView() {
+  return (
+    <VBox
+      style={{
+        gap: 30,
+        backgroundColor: "var(--color-gray-0)",
+        padding: "60px 150px",
+        position: "relative",
+      }}
+    >
+      <Flex style={{ flexDirection: "column", gap: 10, position: "relative" }}>
+        <Alert icon={<EditFilled />}>
+          Your trial has expired. Please purchase to continue.
+        </Alert>
+        <Alert variant="primary" closable={true}>
+          <InfoOutlined /> Your trial has expired. Please purchase to continue.
+        </Alert>
+        <Alert variant="secondary" closable={true}>
+          Your trial has expired. Please purchase to continue.
+        </Alert>
+        <Alert variant="success" closable={true}>
+          Your trial has expired. Please purchase to continue.
+        </Alert>
+        <Alert variant="danger" closable={true}>
+          Your trial has expired. Please purchase to continue.
+        </Alert>
+        <Alert variant="dark" closable={true}>
+          Your trial has expired. Please purchase to continue.
+        </Alert>
+      </Flex>
+    </VBox>
+  );
+}