+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @elf-framework/design-system
 
+## 2.1.21
+
+### Patch Changes
+
+- update Button, dialog, icon, menu, notification, panel, popover, alert style
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@2.1.21
+
 ## 2.1.20
 
 ### Patch Changes